+++ src/main/java/org/c3lang/intellij/C3BuildRunEditor.java
@@ -3,6 +3,7 @@ package org.c3lang.intellij;
 import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory;
 import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.options.SettingsEditor;
+import com.intellij.openapi.ui.TextBrowseFolderListener;
 import com.intellij.openapi.ui.TextFieldWithBrowseButton;
 import com.intellij.util.ui.FormBuilder;
 import org.jetbrains.annotations.NotNull;
@@ -16,13 +17,14 @@ public class C3BuildRunEditor extends SettingsEditor<C3BuildRunConfiguration>
     TextFieldWithBrowseButton workingDirectoryField;
     JTextField argsField;
 
-    public C3BuildRunEditor() {
+    public C3BuildRunEditor()
+    {
         createUIComponents();
 
         panel = FormBuilder.createFormBuilder()
-                .addLabeledComponent("Working directory", workingDirectoryField)
-                .addLabeledComponent("Additional arguments", argsField)
-                .getPanel();
+                           .addLabeledComponent("Working directory", workingDirectoryField)
+                           .addLabeledComponent("Additional arguments", argsField)
+                           .getPanel();
     }
 
     @Override protected void resetEditorFrom(@NotNull C3BuildRunConfiguration configuration)
@@ -30,11 +32,14 @@ public class C3BuildRunEditor extends SettingsEditor<C3BuildRunConfiguration>
         // This function is called each time the run configuration form is shown,
         // i.e. both when its first created and when it's being edited
 
-        if (configuration.getWorkingDirectory().isEmpty()) {
+        if (configuration.getWorkingDirectory().isEmpty())
+        {
             // By default, fill the workingDirectory field with the project's base path
             String projectDirectory = configuration.getProject().getBasePath();
             workingDirectoryField.setText(projectDirectory);
-        } else {
+        }
+        else
+        {
             // Otherwise (when editing the configuration), set its value to the one that was stored
             workingDirectoryField.setText(configuration.getWorkingDirectory());
         }
@@ -43,10 +48,10 @@ public class C3BuildRunEditor extends SettingsEditor<C3BuildRunConfiguration>
         argsField.setText(configuration.getArgs());
     }
 
-    @Override protected void applyEditorTo(@NotNull C3BuildRunConfiguration configuration) throws
-                                                                                           ConfigurationException
+    @Override protected void applyEditorTo(@NotNull C3BuildRunConfiguration configuration) throws ConfigurationException
     {
-        if (workingDirectoryField.getText().isEmpty()) {
+        if (workingDirectoryField.getText().isEmpty())
+        {
             throw new ConfigurationException("You must provide a working directory.");
         }
 
@@ -59,14 +64,13 @@ public class C3BuildRunEditor extends SettingsEditor<C3BuildRunConfiguration>
         return panel;
     }
 
-    private void createUIComponents() {
+    private void createUIComponents()
+    {
         workingDirectoryField = new TextFieldWithBrowseButton();
-        workingDirectoryField.addBrowseFolderListener(
-                "Select Working Directory",
-                null,
-                null,
-                FileChooserDescriptorFactory.createSingleFolderDescriptor()
-        );
+        TextBrowseFolderListener listener =
+                new TextBrowseFolderListener(FileChooserDescriptorFactory.createSingleFolderDescriptor()
+                                                                         .withTitle("Select Working Directory"));
+        workingDirectoryField.addBrowseFolderListener(listener);
 
         argsField = new JTextField();
     }