+++ src/main/java/org/c3lang/intellij/C3CompileRunEditor.java
@@ -6,6 +6,7 @@ import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.options.SettingsEditor;
 import com.intellij.openapi.ui.TextBrowseFolderListener;
 import com.intellij.openapi.ui.TextFieldWithBrowseButton;
+import com.intellij.openapi.util.Comparing;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.ui.FormBuilder;
 import org.jetbrains.annotations.NotNull;
@@ -26,15 +27,16 @@ public class C3CompileRunEditor extends SettingsEditor<C3CompileRunConfiguration
     TextFieldWithBrowseButton workingDirectoryField;
     JTextField argsField;
 
-    public C3CompileRunEditor() {
+    public C3CompileRunEditor()
+    {
         createUIComponents();
 
         panel = FormBuilder.createFormBuilder()
-                .addLabeledComponent("C3 Source file", sourceFileField)
-                .addLabeledComponent("Working directory", workingDirectoryField)
-                .addTooltip("This is the directory where the output executable will be produced.")
-                .addLabeledComponent("Additional arguments", argsField)
-                .getPanel();
+                           .addLabeledComponent("C3 Source file", sourceFileField)
+                           .addLabeledComponent("Working directory", workingDirectoryField)
+                           .addTooltip("This is the directory where the output executable will be produced.")
+                           .addLabeledComponent("Additional arguments", argsField)
+                           .getPanel();
     }
 
     @Override protected void resetEditorFrom(@NotNull C3CompileRunConfiguration configuration)
@@ -42,11 +44,14 @@ public class C3CompileRunEditor extends SettingsEditor<C3CompileRunConfiguration
         // This function is called each time the run configuration form is shown,
         // i.e. both when its first created and when it's being edited
 
-        if (configuration.getWorkingDirectory().isEmpty()) {
+        if (configuration.getWorkingDirectory().isEmpty())
+        {
             // By default, fill the workingDirectory field with the project's base path
             String projectDirectory = configuration.getProject().getBasePath();
             workingDirectoryField.setText(projectDirectory);
-        } else {
+        }
+        else
+        {
             // Otherwise (when editing the configuration), set its value to the one that was stored
             workingDirectoryField.setText(configuration.getWorkingDirectory());
         }
@@ -59,11 +64,13 @@ public class C3CompileRunEditor extends SettingsEditor<C3CompileRunConfiguration
     @Override protected void applyEditorTo(@NotNull C3CompileRunConfiguration configuration) throws
                                                                                              ConfigurationException
     {
-        if (sourceFileField.getText().isEmpty()) {
+        if (sourceFileField.getText().isEmpty())
+        {
             throw new ConfigurationException("You must provide a source file.");
         }
 
-        if (workingDirectoryField.getText().isEmpty()) {
+        if (workingDirectoryField.getText().isEmpty())
+        {
             throw new ConfigurationException("You must provide a working directory.");
         }
 
@@ -80,28 +87,19 @@ public class C3CompileRunEditor extends SettingsEditor<C3CompileRunConfiguration
     private void createUIComponents()
     {
         workingDirectoryField = new TextFieldWithBrowseButton();
-        workingDirectoryField.addBrowseFolderListener(
-                "Select Working Directory",
-                null,
-                null,
-                FileChooserDescriptorFactory.createSingleFolderDescriptor()
-        );
+        TextBrowseFolderListener listener =
+                new TextBrowseFolderListener(FileChooserDescriptorFactory.createSingleFolderDescriptor()
+                                                                         .withTitle("Select Working Directory"));
+        workingDirectoryField.addBrowseFolderListener(listener);
 
         argsField = new JTextField();
 
         sourceFileField = new TextFieldWithBrowseButton();
         FileChooserDescriptor descriptor = new FileChooserDescriptor(true, false, false, false, false, false)
-        {
-            @Override public boolean isFileSelectable(@Nullable VirtualFile file)
-            {
-                if (file == null) return false;
-                if (!file.exists()) return false;
-                if (file.isDirectory()) return false;
-                return Objects.equals(file.getExtension(), "c3i") || Objects.equals(file.getExtension(), "c3");
-            }
-        };
-        sourceFileField.addBrowseFolderListener(
-                new TextBrowseFolderListener(descriptor.withTitle("Select C3 Source File"))
-        );
+                .withFileFilter((VirtualFile file) ->
+                                        Comparing.equal(file.getExtension(), "c3i", true)
+                                        || Comparing.equal(file.getExtension(), "c3", true));
+        sourceFileField.addBrowseFolderListener(new TextBrowseFolderListener(descriptor.withTitle(
+                "Select C3 Source File")));
     }
 }